fix is_prime_prog for bigger number

+++ is_prime_prog/main.cpp
@@ -3,7 +3,7 @@
 
 int main()
 {
-    int i, n;
+    long long i, n;
 
     while (std::cin >> n)
     {
